refactor(playground-ui): split sidebar active helper
