+++ packages/playground-ui/src/ds/components/MainSidebar/main-sidebar.tsx
@@ -18,7 +18,7 @@ export { type NavLink } from './main-sidebar-nav-link';
 export { type NavSection } from './main-sidebar-nav-section';
 export { MainSidebarTrigger } from './main-sidebar-trigger';
 export { MainSidebarMobileTrigger } from './main-sidebar-mobile-trigger';
-export { getIsLinkActive } from './main-sidebar-sections';
+export { getIsLinkActive } from './main-sidebar-link-active';
 
 export const MainSidebar = Object.assign(MainSidebarRoot, {
   Bottom: MainSidebarBottom,
